refactor(rcml): delete brand-elements module; inline its remaining users into the CLI
Once every vendor template migrated to the XML / createEmailTemplate
pattern, brand-elements was an orphan with only a single consumer left
(the validate-rcml CLI showcase). Move the 12 builders that command
still uses into a sibling validate-rcml-builders.ts and drop the rcml
file along with its public re-export.
