+++ packages/rcml/src/email/index.ts
@@ -21,5 +21,4 @@ export * from './theme-defaults.js'
 export * from './create-theme.js'
 export * from './apply-theme.js'
 export * from './get-theme.js'
-export * from './brand-elements.js'
 export * from './create-email-template.js'
